+++ Spoony-iOS/Spoony-iOS/Source/Feature/Home/Mypage/AttendanceInfoSheetView.swift
@@ -23,7 +23,7 @@ struct AttendanceInfoSheetView: View {
                             Image("VSpoon")
                                 .frame(width: 100, height: 100)
                          
-                        VStack(alignment: .leading, spacing: 12) {
+                        VStack(alignment: .leading, spacing: 8.adjusted) {
                             Text("오늘의 스푼 받기")
                                 .customFont(.title2)
                                 .foregroundColor(.white)
@@ -69,28 +69,6 @@ struct AttendanceInfoSheetView: View {
         }
         .presentationDetents([.height(900)])
         .presentationDragIndicator(.hidden)
-    }
-}
-
-struct StampCategoryRow: View {
-    let title: String
-    let count: String
-    
-    var body: some View {
-        HStack(spacing: 2) {
-            Text(title)
-                .customFont(.body1m)
-                .foregroundColor(.gray700)
-                .frame(width: 130.adjusted, height: 47.adjustedH)
-                .background(Color.gray100)
-            
-            Text(count)
-                .customFont(.body1m)
-                .foregroundColor(.gray700)
-                .frame(width: 203.adjusted, height: 47.adjustedH)
-                .background(Color.gray0)
-        }
-        .padding(.horizontal, 20)
     }
 }
